pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coldshot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "httpx>=0.28",
    "pydantic>=2.10",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["sumble", "pipeline", "mailer", "recorder"]

[tool.ruff]
target-version = "py312"
